🚚 Move CPU/IO/ACPI from HAL.X64/ARM64 into Core.X64/ARM64
