+++ src/Cosmos.Kernel.Core.ARM64/Cpu/GIC.cs
@@ -2,9 +2,8 @@
 
 using Cosmos.Kernel.Boot.Limine;
 using Cosmos.Kernel.Core.IO;
-using Cosmos.Kernel.HAL.ARM64;
 
-namespace Cosmos.Kernel.HAL.ARM64.Cpu;
+namespace Cosmos.Kernel.Core.ARM64.Cpu;
 
 /// <summary>
 /// Unified GIC facade that auto-detects GICv2 or GICv3 at runtime
@@ -75,7 +74,7 @@ public static class GIC
     public static unsafe void Initialize()
     {
         // Priority 1: Try ACPI MADT (parsed by C code in kmain via acpi_early_init)
-        var acpiGic = Acpi.GetGicInfo();
+        var acpiGic = AcpiGic.GetGicInfo();
         if (acpiGic != null && acpiGic->Found != 0)
         {
             _distBase = acpiGic->DistBase;
