+++ src/Cosmos.Kernel.Core.ARM64/Cpu/GICv3.cs
@@ -4,7 +4,7 @@ using System.Runtime.CompilerServices;
 using Cosmos.Kernel.Core.ARM64.Bridge;
 using Cosmos.Kernel.Core.IO;
 
-namespace Cosmos.Kernel.HAL.ARM64.Cpu;
+namespace Cosmos.Kernel.Core.ARM64.Cpu;
 
 /// <summary>
 /// ARM Generic Interrupt Controller v3 (GICv3) implementation.
