+++ src/Cosmos.Kernel.Core.X64/Cpu/X64CpuOps.cs
@@ -1,9 +1,8 @@
 using Cosmos.Kernel.Core.CPU;
 using Cosmos.Kernel.Core.IO;
 using Cosmos.Kernel.Core.X64.Bridge;
-using Cosmos.Kernel.HAL.Interfaces;
 
-namespace Cosmos.Kernel.HAL.X64;
+namespace Cosmos.Kernel.Core.X64.Cpu;
 
 public class X64CpuOps : ICpuOps
 {
