+++ src/Cosmos.Kernel.HAL.ARM64/Devices/Clock/RTC.cs
@@ -3,7 +3,7 @@
 using Cosmos.Kernel.Boot.Limine;
 using Cosmos.Kernel.Core;
 using Cosmos.Kernel.Core.IO;
-using Cosmos.Kernel.HAL.ARM64.Cpu;
+using Cosmos.Kernel.Core.ARM64.Cpu;
 using Cosmos.Kernel.HAL.ARM64.Devices.Timer;
 using Cosmos.Kernel.HAL.Devices;
 using Cosmos.Kernel.HAL.Devices.Clock;
