+++ src/Cosmos.Kernel.HAL.ARM64/Devices/Input/VirtioKeyboard.cs
@@ -3,7 +3,7 @@
 using System.Runtime.InteropServices;
 using Cosmos.Kernel.Core.IO;
 using Cosmos.Kernel.Core.Memory;
-using Cosmos.Kernel.HAL.ARM64.Cpu;
+using Cosmos.Kernel.Core.ARM64.Cpu;
 using Cosmos.Kernel.HAL.ARM64.Devices.Virtio;
 using Cosmos.Kernel.Core.CPU;
 using Cosmos.Kernel.HAL.Devices.Input;
