+++ src/Cosmos.Kernel.HAL.X64/Devices/Input/PS2Controller.cs
@@ -312,7 +312,7 @@ public class PS2Controller : Device
         byte irq = (byte)(port == 1 ? 1 : 12);
 
         // Unmask at I/O APIC
-        Cosmos.Kernel.HAL.X64.Cpu.ApicManager.UnmaskIrq(irq);
+        Cosmos.Kernel.Core.X64.Cpu.ApicManager.UnmaskIrq(irq);
     }
 
     /// <summary>
