+++ tests/Kernels/Cosmos.Kernel.Tests.Timer/Kernel.cs
@@ -6,7 +6,7 @@ using Cosmos.TestRunner.Framework;
 using Sys = Cosmos.Kernel.System;
 using TR = Cosmos.TestRunner.Framework.TestRunner;
 #if ARCH_X64
-using Cosmos.Kernel.HAL.X64.Cpu;
+using Cosmos.Kernel.Core.X64.Cpu;
 using Cosmos.Kernel.HAL.X64.Devices.Clock;
 using Cosmos.Kernel.HAL.X64.Devices.Timer;
 #else
